requirements.txt
amqp==2.4.1
billiard==3.5.0.5
celery==4.2.1
construct==2.5.3
coverage==4.5.2
Django==2.1.3
django-haystack==2.8.1
future==0.17.1
imageio==2.5.0
imageio-ffmpeg==0.2.0
jieba==0.39
kombu==4.3.0
numpy==1.16.1
pefile==2018.8.8
Pillow==5.4.1
python-ptrace==0.9.3
pytz==2018.7
redis==3.2.0
six==1.12.0
vine==1.2.0
Whoosh==2.7.4
